Fisiese movement van pieces

+++ Downloads/Jaar3/Werk31/REII 313/Skaak_V2/chesspiece.cpp
@@ -24,7 +24,7 @@ QList<Position> ChessPiece::possibleMoves()
     return moves;
 }
 
-void ChessPiece::movePieceTo(int xx, int yy)
+void ChessPiece::movePieceTo(float xx, float yy)
 {
     // Helper function to move chess piece according to board[][] coordinates.
